demos/monitors: remove iostream, remove warnings

+++ demos/crossroad_camera_demo/main.cpp
@@ -607,7 +607,7 @@ int main(int argc, char *argv[]) {
         std::cout << std::endl;
 
         cv::Size graphSize{static_cast<int>(cap.get(cv::CAP_PROP_FRAME_WIDTH) / 4), 60};
-        Presenter presenter(FLAGS_u, cap.get(cv::CAP_PROP_FRAME_HEIGHT) - graphSize.height - 10, graphSize);
+        Presenter presenter(FLAGS_u, static_cast<int>(cap.get(cv::CAP_PROP_FRAME_HEIGHT)) - graphSize.height - 10, graphSize);
 
         do {
             // get and enqueue the next frame (in case of video)
