+++ demos/human_pose_estimation_demo/main.cpp
@@ -76,7 +76,7 @@ int main(int argc, char* argv[]) {
         std::cout << std::endl;
 
         cv::Size graphSize{static_cast<int>(cap.get(cv::CAP_PROP_FRAME_WIDTH) / 4), 60};
-        Presenter presenter(FLAGS_u, cap.get(cv::CAP_PROP_FRAME_HEIGHT) - graphSize.height - 10, graphSize);
+        Presenter presenter(FLAGS_u, static_cast<int>(cap.get(cv::CAP_PROP_FRAME_HEIGHT)) - graphSize.height - 10, graphSize);
 
         do {
             double t1 = static_cast<double>(cv::getTickCount());
